Specify identical populations
This ensures that the same dates get used across both approaches (ehrQL vs cohortextractor). Previously we used the "index_date" 2022-09-25 specified in the `project.yaml`. Looking at the SQL code generated by the cohortextractor implementation we noticed that the start date 2020-04-26 was used instead. We expect this explains differences in the variables `care_home_code` and `care_home_tpp`.

We now also use the ons death for the variable `has_died`.